Fill out README
I removed the "name" section as I don't really think it's useful. But
for reference, I will explain it in this commit message: "Ogrim" is the
original name of "Dung Defender", which is a dung beetle. Dung beetles
build stuff from dung. And XML is considered by many to be kind of
shitty... It's still required and useful in places, hence this library.

https://hollowknight.fandom.com/wiki/Dung_Defender